#549 organize C++ builders in separate files
