+++ include/flecs/addons/cpp/flecs.hpp
@@ -224,12 +224,13 @@ static const flecs::entity_t Throw = EcsThrow;
 #include "world.hpp"
 #include "id.hpp"
 #include "entity_view.hpp"
+#include "ref.hpp"
 #include "entity.hpp"
 #include "component.hpp"
 #include "invoker.hpp"
-#include "builder.hpp"
 #include "type.hpp"
 #include "module.hpp"
+#include "term.hpp"
 #include "filter.hpp"
 #include "snapshot.hpp"
 #include "filter_iterator.hpp"
