+++ include/flecs/addons/cpp/ref.hpp
@@ -1,0 +1,52 @@
+namespace flecs
+{
+
+template <typename T>
+class ref {
+public:
+    ref()
+        : m_world( nullptr )
+        , m_entity( 0 )
+        , m_ref() { }
+
+    ref(world_t *world, entity_t entity) 
+        : m_world( world )
+        , m_entity( entity )
+        , m_ref() 
+    {
+        auto comp_id = _::cpp_type<T>::id(world);
+
+        ecs_assert(_::cpp_type<T>::size() != 0, 
+                ECS_INVALID_PARAMETER, NULL);
+
+        ecs_get_ref_w_id(
+            m_world, &m_ref, m_entity, comp_id);
+    }
+
+    const T* operator->() {
+        const T* result = static_cast<const T*>(ecs_get_ref_w_id(
+            m_world, &m_ref, m_entity, _::cpp_type<T>::id(m_world)));
+
+        ecs_assert(result != NULL, ECS_INVALID_PARAMETER, NULL);
+
+        return result;
+    }
+
+    const T* get() {
+        if (m_entity) {
+            ecs_get_ref_w_id(
+                m_world, &m_ref, m_entity, _::cpp_type<T>::id(m_world));    
+        }
+
+        return static_cast<T*>(m_ref.ptr);
+    }
+
+    flecs::entity entity() const;
+
+private:
+    world_t *m_world;
+    entity_t m_entity;
+    flecs::ref_t m_ref;
+};
+
+}
